server/requirements.txt

Django==1.11.29
djangorestframework==3.9.4
django-cors-headers==3.0.2